Adds a bandolier to the warden's locker (#94130)
## About The Pull Request

PR title

## Why It's Good For The Game

Warden is the weapons guy and even starts with a super cool unique
shotgun. Bandoliers are roundstart items (can be bought from bartender's
vendor when they arent using it or arent preset, can be crafted by
killing five monkeys which there are plenty of) so this just removes the
hastle of having to go get it.

## Changelog
:cl:
balance: The warden starts with a bandolier in their locker
/:cl:

Co-authored-by: John Doe <markkavalerov87@gmail.com>